+++ CHANGELOG.md
@@ -1,9 +1,19 @@
-# Change Log
+# Changelog
 
-All notable changes to the "anglr-formatter-vscode" extension will be documented in this file.
+## Version 0.2.0
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+- updated version of `anglr-formatter`
+- *formatter* configuration
+    - `anglr.callExpressionArgumentsFormatter.enable` - Enable/disalbe automatic Anglr call expression arguments formatter.
+    - `anglr.constructorParameterFormatter.enable` - Enable/disalbe automatic Anglr constructor formatter.
+    - `anglr.decoratorArgumentsFormatter.enable` - Enable/disalbe automatic Anglr decorator arguments formatter.
+    - `anglr.importFormatter.enable` - Enable/disalbe automatic Anglr import formatter.
+    - `anglr.localVersion.enable` - Enable/disalbe automatic use of local node_modules version.
 
-## [Unreleased]
+## Version 0.1.0
 
-- Initial release
+- first version of formatter, changes whole files
+- allows disabling of formatter `anglr.format.enable`, requires restart
+- *formatter* configuration
+    - `anglr.importReorder.enable` - Enable/disalbe automatic Anglr formatter import reordering.
+